testing/20140814.html

<!DOCTYPE html>
<title>IRC logs: w3c / #testing / 20140814</title>
<meta name="viewport" content="width=device-width">
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; word-wrap: break-word; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<script src="/irc-logs/s.js"></script>
<h1 id="w3c/testing.20140814.log"><a href="/irc-logs/">/irc-logs</a> / w3c / <a href="irc://irc.w3.org:6665/testing">#testing</a> / 2014-08-14 / <a href="#end">end</a></h1>
<p>Options: 
<ol id="lines">
 <li id="l-1"><a href="#l-1">#</a> Session Start: Thu Aug 14 00:00:01 2014</li>
 <li id="l-2"><a href="#l-2">#</a> Session Ident: #testing</li>
 <li id="l-3" _a="MikeSmith"><a href="#l-3">#</a> [00:31] &lt;MikeSmith&gt; shotdown++</li>
 <li id="l-4" _a="jgraham" _r="ato"><a href="#l-4">#</a> [00:34] &lt;jgraham&gt; ato: Thanks</li>
 <li id="l-5" _a="ato" _r="jgraham"><a href="#l-5">#</a> [00:35] &lt;ato&gt; jgraham: Freudian slip? (-;</li>
 <li id="l-6"><a href="#l-6">#</a> [00:42] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Client closed connection)</li>
 <li id="l-7"><a href="#l-7">#</a> [01:23] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-8"><a href="#l-8">#</a> [01:51] * Quits: lmcliste_ (~lmclister@public.cloak) (&quot;&quot;)</li>
 <li id="l-9"><a href="#l-9">#</a> [02:02] * Joins: glenn_ (~gadams@public.cloak)</li>
 <li id="l-10"><a href="#l-10">#</a> [02:04] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-11"><a href="#l-11">#</a> [02:14] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-12"><a href="#l-12">#</a> [02:18] * Joins: glenn__ (~gadams@public.cloak)</li>
 <li id="l-13"><a href="#l-13">#</a> [02:19] * Quits: glenn_ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-14"><a href="#l-14">#</a> [02:21] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-15"><a href="#l-15">#</a> [02:23] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-16"><a href="#l-16">#</a> [02:30] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-17"><a href="#l-17">#</a> [02:33] * Quits: glenn__ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-18"><a href="#l-18">#</a> [02:38] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-19"><a href="#l-19">#</a> [02:39] * Joins: glenn_ (~gadams@public.cloak)</li>
 <li id="l-20"><a href="#l-20">#</a> [02:40] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-21" _a="heycam|away"><a href="#l-21">#</a> [02:42] * heycam|away is now known as heycam</li>
 <li id="l-22"><a href="#l-22">#</a> [02:58] * Quits: ptressel (~chatzilla@public.cloak) (&quot;bbl&quot;)</li>
 <li id="l-23"><a href="#l-23">#</a> [03:22] * Quits: ArtB (~abarsto@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-24" _a="heycam"><a href="#l-24">#</a> [04:05] * heycam is now known as heycam|away</li>
 <li id="l-25"><a href="#l-25">#</a> [04:26] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-26"><a href="#l-26">#</a> [05:55] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-27"><a href="#l-27">#</a> [06:00] * Quits: glenn_ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-28" _a="heycam|away"><a href="#l-28">#</a> [06:10] * heycam|away is now known as heycam</li>
 <li id="l-29"><a href="#l-29">#</a> [06:11] * Quits: lmclister (~lmclister@public.cloak) (&quot;&quot;)</li>
 <li id="l-30" _a="heycam"><a href="#l-30">#</a> [07:01] * heycam is now known as heycam|away</li>
 <li id="l-31"><a href="#l-31">#</a> [07:34] * Joins: ptressel (~chatzilla@public.cloak)</li>
 <li id="l-32"><a href="#l-32">#</a> [08:06] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-33"><a href="#l-33">#</a> [08:13] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-34" _a="heycam|away"><a href="#l-34">#</a> [08:56] * heycam|away is now known as heycam</li>
 <li id="l-35"><a href="#l-35">#</a> [08:59] * Quits: lmclister (~lmclister@public.cloak) (&quot;&quot;)</li>
 <li id="l-36"><a href="#l-36">#</a> [09:11] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-37"><a href="#l-37">#</a> [09:22] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-38"><a href="#l-38">#</a> [09:22] -gitbot:#testing- [web-platform-tests] isairz opened pull request #1173: Check disabled-video at next frame (master...track-disable) <a href="https://github.com/w3c/web-platform-tests/pull/1173">https://github.com/w3c/web-platform-tests/pull/1173</a></li>
 <li id="l-39"><a href="#l-39">#</a> [09:22] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-40"><a href="#l-40">#</a> [09:25] * Joins: dom (dom@public.cloak)</li>
 <li id="l-41"><a href="#l-41">#</a> [09:45] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-42"><a href="#l-42">#</a> [09:51] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Client closed connection)</li>
 <li id="l-43"><a href="#l-43">#</a> [09:51] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-44"><a href="#l-44">#</a> [09:58] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-45"><a href="#l-45">#</a> [10:09] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-46"><a href="#l-46">#</a> [10:30] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-47"><a href="#l-47">#</a> [10:30] -gitbot:#testing- [web-platform-tests] jgraham pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/f04e6100c53ba1a4abf22ee17ce2d9cd79d006db">https://github.com/w3c/web-platform-tests/commit/f04e6100c53ba1a4abf22ee17ce2d9cd79d006db</a></li>
 <li id="l-48"><a href="#l-48">#</a> [10:30] -gitbot:#testing- web-platform-tests/master f04e610 James Graham: Update to latest wptserve.</li>
 <li id="l-49"><a href="#l-49">#</a> [10:30] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-50"><a href="#l-50">#</a> [10:33] * Quits: ptressel (~chatzilla@public.cloak) (&quot;zzz&quot;)</li>
 <li id="l-51"><a href="#l-51">#</a> [10:38] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-52"><a href="#l-52">#</a> [10:40] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-53" _a="plinss"><a href="#l-53">#</a> [10:44] * plinss is now known as plinss_away</li>
 <li id="l-54" _a="plinss_away"><a href="#l-54">#</a> [10:48] * plinss_away is now known as plinss</li>
 <li id="l-55"><a href="#l-55">#</a> [10:52] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-56"><a href="#l-56">#</a> [11:11] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-57"><a href="#l-57">#</a> [11:11] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #1174: Whitelist gifs from the CR-at-EOL lint check. (master...cr-gif) <a href="https://github.com/w3c/web-platform-tests/pull/1174">https://github.com/w3c/web-platform-tests/pull/1174</a></li>
 <li id="l-58"><a href="#l-58">#</a> [11:11] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-59"><a href="#l-59">#</a> [11:11] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-60"><a href="#l-60">#</a> [11:11] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #1174: Whitelist gifs from the CR-at-EOL lint check. (master...cr-gif) <a href="https://github.com/w3c/web-platform-tests/pull/1174">https://github.com/w3c/web-platform-tests/pull/1174</a></li>
 <li id="l-61"><a href="#l-61">#</a> [11:11] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-62"><a href="#l-62">#</a> [11:11] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-63"><a href="#l-63">#</a> [11:11] -gitbot:#testing- [web-platform-tests] jgraham pushed 2 new commits to master: <a href="https://github.com/w3c/web-platform-tests/compare/f04e6100c53b...c711b6fb2912">https://github.com/w3c/web-platform-tests/compare/f04e6100c53b...c711b6fb2912</a></li>
 <li id="l-64"><a href="#l-64">#</a> [11:11] -gitbot:#testing- web-platform-tests/master a48acc4 Ms2ger: Whitelist gifs from the CR-at-EOL lint check.</li>
 <li id="l-65"><a href="#l-65">#</a> [11:11] -gitbot:#testing- web-platform-tests/master c711b6f jgraham: Merge pull request #1174 from Ms2ger/cr-gif...</li>
 <li id="l-66"><a href="#l-66">#</a> [11:11] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-67" _a="Ms2ger" _r="jgraham"><a href="#l-67">#</a> [11:12] &lt;Ms2ger&gt; jgraham, r?... Oh</li>
 <li id="l-68"><a href="#l-68">#</a> [11:13] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-69"><a href="#l-69">#</a> [11:13] -gitbot:#testing- [web-platform-tests] Ms2ger pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/aec973a22ccf8af0668b543adfbef2ebd533bb3b">https://github.com/w3c/web-platform-tests/commit/aec973a22ccf8af0668b543adfbef2ebd533bb3b</a></li>
 <li id="l-70"><a href="#l-70">#</a> [11:13] -gitbot:#testing- web-platform-tests/master aec973a Ms2ger: Merge pull request #1115 from w3c/hallvors/gzipfix2...</li>
 <li id="l-71"><a href="#l-71">#</a> [11:13] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-72"><a href="#l-72">#</a> [11:13] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-73"><a href="#l-73">#</a> [11:13] -gitbot:#testing- [web-platform-tests] Ms2ger deleted hallvors/gzipfix2 at 224c2ae: <a href="https://github.com/w3c/web-platform-tests/commit/224c2ae">https://github.com/w3c/web-platform-tests/commit/224c2ae</a></li>
 <li id="l-74"><a href="#l-74">#</a> [11:13] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-75"><a href="#l-75">#</a> [11:28] * Joins: rhauck (~rhauck@public.cloak)</li>
 <li id="l-76"><a href="#l-76">#</a> [11:31] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-77"><a href="#l-77">#</a> [11:35] * Quits: rhauck (~rhauck@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-78"><a href="#l-78">#</a> [11:38] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-79"><a href="#l-79">#</a> [11:40] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-80"><a href="#l-80">#</a> [11:40] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #864: Mulder21c (master...mulder21c) <a href="https://github.com/w3c/web-platform-tests/pull/864">https://github.com/w3c/web-platform-tests/pull/864</a></li>
 <li id="l-81"><a href="#l-81">#</a> [11:40] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-82" _a="MikeSmith"><a href="#l-82">#</a> [11:44] &lt;MikeSmith&gt; we need a &quot;Nice close!&quot; emoticon to go along with the &quot;Nice merge!&quot; love emoticon</li>
 <li id="l-83" _a="MikeSmith"><a href="#l-83">#</a> [11:45] &lt;MikeSmith&gt; down to 35 stale-awaiting-submitter-response PRs <a href="https://github.com/w3c/web-platform-tests/labels/stale-awaiting-submitter-response">https://github.com/w3c/web-platform-tests/labels/stale-awaiting-submitter-response</a></li>
 <li id="l-84" _a="MikeSmith"><a href="#l-84">#</a> [11:45] &lt;MikeSmith&gt; and 19 stale-awaiting-review <a href="https://github.com/w3c/web-platform-tests/labels/stale-awaiting-review">https://github.com/w3c/web-platform-tests/labels/stale-awaiting-review</a></li>
 <li id="l-85" _a="ato"><a href="#l-85">#</a> [11:49] &lt;ato&gt; I've pinged mdas about getting push rights to her branch.</li>
 <li id="l-86" _a="Ms2ger"><a href="#l-86">#</a> [11:51] * Ms2ger needed an excuse not to do his actual work</li>
 <li id="l-87"><a href="#l-87">#</a> [11:57] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-88"><a href="#l-88">#</a> [11:57] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #871: addCue and removeCue Test Source (master...yongsu9/flexbox-flex-direction-prop) <a href="https://github.com/w3c/web-platform-tests/pull/871">https://github.com/w3c/web-platform-tests/pull/871</a></li>
 <li id="l-89"><a href="#l-89">#</a> [11:57] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-90" _a="MikeSmith" _r="ato"><a href="#l-90">#</a> [12:29] &lt;MikeSmith&gt; ato: cool</li>
 <li id="l-91" _a="Ms2ger"><a href="#l-91">#</a> [12:30] &lt;Ms2ger&gt; Upcoming Events</li>
 <li id="l-92" _a="Ms2ger"><a href="#l-92">#</a> [12:30] &lt;Ms2ger&gt; August 3, 2014 - Portland, Oregon.</li>
 <li id="l-93" _a="MikeSmith" _r="Ms2ger"><a href="#l-93">#</a> [12:30] &lt;MikeSmith&gt; Ms2ger: yay for not doing your actual work but closing out more PRs instead. Lead by example!</li>
 <li id="l-94" _a="Ms2ger"><a href="#l-94">#</a> [12:31] &lt;Ms2ger&gt; We used to have a graph of outstanding PRs on ttwf, no?</li>
 <li id="l-95" _a="MikeSmith" _r="Ms2ger"><a href="#l-95">#</a> [12:31] &lt;MikeSmith&gt; Ms2ger: just change the heading to Events that Either Were Upcoming in the Past or that Might actually Be Coming up in the Future</li>
 <li id="l-96" _a="MikeSmith"><a href="#l-96">#</a> [12:31] &lt;MikeSmith&gt; yeah we had a graph</li>
 <li id="l-97" _a="MikeSmith"><a href="#l-97">#</a> [12:31] &lt;MikeSmith&gt; dunno what happened to it</li>
 <li id="l-98" _a="MikeSmith"><a href="#l-98">#</a> [12:31] &lt;MikeSmith&gt; ask plh he likes making graphs</li>
 <li id="l-99" _a="Ms2ger"><a href="#l-99">#</a> [12:33] &lt;Ms2ger&gt; <a href="http://testthewebforward.org/dashboard/">http://testthewebforward.org/dashboard/</a></li>
 <li id="l-100" _a="Ms2ger" _r="MikeSmith"><a href="#l-100">#</a> [12:35] &lt;Ms2ger&gt; MikeSmith, since you're in a yay mood, r? <a href="https://critic.hoppipolla.co.uk/r/2323">https://critic.hoppipolla.co.uk/r/2323</a></li>
 <li id="l-101" _a="MikeSmith"><a href="#l-101">#</a> [12:35] * MikeSmith looks</li>
 <li id="l-102" _a="Ms2ger"><a href="#l-102">#</a> [12:35] &lt;Ms2ger&gt; Shouldn't be too hard, it's mostly moving around</li>
 <li id="l-103" _a="MikeSmith" _r="Ms2ger"><a href="#l-103">#</a> [12:36] &lt;MikeSmith&gt; Ms2ger: yeah will do it right now</li>
 <li id="l-104" _a="Ms2ger"><a href="#l-104">#</a> [12:36] &lt;Ms2ger&gt; &lt;3</li>
 <li id="l-105" _a="jgraham"><a href="#l-105">#</a> [12:36] &lt;jgraham&gt; testtwf.org needs a lot of love</li>
 <li id="l-106" _a="jgraham"><a href="#l-106">#</a> [12:36] &lt;jgraham&gt; A whole lotta love in fact</li>
 <li id="l-107" _a="Ms2ger"><a href="#l-107">#</a> [12:36] * Ms2ger wonders how much shipping chocolate to .jp would cost</li>
 <li id="l-108" _a="MikeSmith"><a href="#l-108">#</a> [12:37] &lt;MikeSmith&gt; hmm whitespace in those files for <a href="https://critic.hoppipolla.co.uk/r/2323">https://critic.hoppipolla.co.uk/r/2323</a></li>
 <li id="l-109" _a="MikeSmith" _r="Ms2ger"><a href="#l-109">#</a> [12:37] &lt;MikeSmith&gt; Ms2ger: you want me to flag the whitespace, or we don't care for these?</li>
 <li id="l-110" _a="Ms2ger"><a href="#l-110">#</a> [12:37] &lt;Ms2ger&gt; I thought I fixed it?</li>
 <li id="l-111" _a="Ms2ger"><a href="#l-111">#</a> [12:37] &lt;Ms2ger&gt; Hrm</li>
 <li id="l-112" _a="Ms2ger" _r="MikeSmith"><a href="#l-112">#</a> [12:38] &lt;Ms2ger&gt; MikeSmith, flag, please</li>
 <li id="l-113" _a="MikeSmith" _r="Ms2ger"><a href="#l-113">#</a> [12:38] &lt;MikeSmith&gt; Ms2ger: yeah I still see red in <a href="https://critic.hoppipolla.co.uk/2d4d4fe2?review=2323">https://critic.hoppipolla.co.uk/2d4d4fe2?review=2323</a></li>
 <li id="l-114" _a="MikeSmith"><a href="#l-114">#</a> [12:38] &lt;MikeSmith&gt; ok will flag it</li>
 <li id="l-115"><a href="#l-115">#</a> [12:39] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-116"><a href="#l-116">#</a> [12:39] -gitbot:#testing- [web-platform-tests] zcorpan pushed 5 new commits to master: <a href="https://github.com/w3c/web-platform-tests/compare/aec973a22ccf...3762fd15a9db">https://github.com/w3c/web-platform-tests/compare/aec973a22ccf...3762fd15a9db</a></li>
 <li id="l-117"><a href="#l-117">#</a> [12:39] -gitbot:#testing- web-platform-tests/master 34cab8b Philip Jägenstedt: Add a manual test for providing a stable state in alert/confirm/prompt...</li>
 <li id="l-118"><a href="#l-118">#</a> [12:39] -gitbot:#testing- web-platform-tests/master 8e5a5bd Philip Jägenstedt: Update the stable state in dialogs test to match spec change...</li>
 <li id="l-119"><a href="#l-119">#</a> [12:39] -gitbot:#testing- web-platform-tests/master 849effb Philip Jägenstedt: Add a manual test for providing a stable state in print()...</li>
 <li id="l-120"><a href="#l-120">#</a> [12:39] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-121" _a="MikeSmith" _r="Ms2ger"><a href="#l-121">#</a> [12:40] &lt;MikeSmith&gt; Ms2ger: what's with the .xht extension? we want to keep that as-is?</li>
 <li id="l-122" _a="Ms2ger"><a href="#l-122">#</a> [12:41] &lt;Ms2ger&gt; As opposed to .xhtml? I don't care either way</li>
 <li id="l-123" _a="MikeSmith"><a href="#l-123">#</a> [12:43] &lt;MikeSmith&gt; oh if you don't care I don't</li>
 <li id="l-124" _a="MikeSmith" _r="Ms2ger"><a href="#l-124">#</a> [12:45] &lt;MikeSmith&gt; Ms2ger: ok done</li>
 <li id="l-125" _a="Ms2ger"><a href="#l-125">#</a> [12:46] &lt;Ms2ger&gt; Thanks</li>
 <li id="l-126" _a="MikeSmith"><a href="#l-126">#</a> [12:47] &lt;MikeSmith&gt; nice to see <a href="https://travis-ci.org/servo/servo">https://travis-ci.org/servo/servo</a></li>
 <li id="l-127" _a="jgraham"><a href="#l-127">#</a> [12:51] &lt;jgraham&gt; I wish I felt more enthusiastic about doing an overhaul of the website. I know what needs to be done, but there's lots of tedious to get there</li>
 <li id="l-128" _a="Ms2ger" _r="MikeSmith"><a href="#l-128">#</a> [12:51] &lt;Ms2ger&gt; MikeSmith, r? <a href="https://critic.hoppipolla.co.uk/r/2323">https://critic.hoppipolla.co.uk/r/2323</a></li>
 <li id="l-129" _a="jgraham"><a href="#l-129">#</a> [12:52] &lt;jgraham&gt; (done)</li>
 <li id="l-130" _a="jgraham"><a href="#l-130">#</a> [12:52] &lt;jgraham&gt; Would have been nice if you'd cleaned up the tabs at the same time</li>
 <li id="l-131" _a="Ms2ger"><a href="#l-131">#</a> [12:53] &lt;Ms2ger&gt; I guess I can do that too</li>
 <li id="l-132" _a="Ms2ger" _r="jgraham"><a href="#l-132">#</a> [12:54] &lt;Ms2ger&gt; jgraham, r? :)</li>
 <li id="l-133" _a="jgraham"><a href="#l-133">#</a> [12:55] &lt;jgraham&gt; r+</li>
 <li id="l-134" _a="Ms2ger"><a href="#l-134">#</a> [12:56] &lt;Ms2ger&gt; Ta</li>
 <li id="l-135"><a href="#l-135">#</a> [12:57] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-136"><a href="#l-136">#</a> [12:57] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #1167: Cleanup the querySelector{,All} tests. (master...selectors) <a href="https://github.com/w3c/web-platform-tests/pull/1167">https://github.com/w3c/web-platform-tests/pull/1167</a></li>
 <li id="l-137"><a href="#l-137">#</a> [12:57] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-138" _a="Ms2ger" _r="Woo"><a href="#l-138">#</a> [12:58] &lt;Ms2ger&gt; Woo, down to 118</li>
 <li id="l-139"><a href="#l-139">#</a> [13:03] * Joins: abarsto (~abarsto@public.cloak)</li>
 <li id="l-140" _a="abarsto"><a href="#l-140">#</a> [13:03] * abarsto is now known as ArtB</li>
 <li id="l-141"><a href="#l-141">#</a> [13:42] * Quits: wilhelm (~wilhelm@public.cloak) (&quot;Changing server&quot;)</li>
 <li id="l-142"><a href="#l-142">#</a> [13:57] * Quits: MK (kawamori@public.cloak) (&quot;Leaving...&quot;)</li>
 <li id="l-143"><a href="#l-143">#</a> [13:59] * Quits: Lachy (~Lachy@public.cloak) (&quot;Textual IRC Client: www.textualapp.com&quot;)</li>
 <li id="l-144"><a href="#l-144">#</a> [13:59] * Joins: plh (plehegar@public.cloak)</li>
 <li id="l-145" _a="MikeSmith"><a href="#l-145">#</a> [13:59] &lt;MikeSmith&gt; excellemente</li>
 <li id="l-146" _a="Ms2ger"><a href="#l-146">#</a> [14:00] &lt;Ms2ger&gt; Now we need plh... :)</li>
 <li id="l-147"><a href="#l-147">#</a> [14:01] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-148" _a="MikeSmith" _r="plh"><a href="#l-148">#</a> [14:03] &lt;MikeSmith&gt; plh: down to only 118 open PRs thanks to Ms2ger blowing of his day job and working on wpt stuff instead</li>
 <li id="l-149" _a="MikeSmith" _r="plh"><a href="#l-149">#</a> [14:04] &lt;MikeSmith&gt; plh: I nominate Ms2ger for wpt Employee of the Month</li>
 <li id="l-150" _a="Ms2ger"><a href="#l-150">#</a> [14:04] &lt;Ms2ger&gt; Next up: plh ditching his day job for <a href="https://github.com/w3c/web-platform-tests/pull/1077">https://github.com/w3c/web-platform-tests/pull/1077</a></li>
 <li id="l-151" _a="MikeSmith" _r="plh"><a href="#l-151">#</a> [14:06] &lt;MikeSmith&gt; plh: please get your priorities straight and follow Employee of the Month Ms2ger's lead and expedite that PR</li>
 <li id="l-152" _a="Ms2ger"><a href="#l-152">#</a> [14:06] * Ms2ger looks for something else to make MikeSmith do</li>
 <li id="l-153" _a="plh"><a href="#l-153">#</a> [14:07] * plh tries to wake up</li>
 <li id="l-154" _a="Ms2ger"><a href="#l-154">#</a> [14:08] * Ms2ger passes the coffee</li>
 <li id="l-155"><a href="#l-155">#</a> [14:08] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-156"><a href="#l-156">#</a> [14:08] -gitbot:#testing- [web-platform-tests] dontcallmedom closed pull request #1173: Check disabled-video at next frame (master...track-disable) <a href="https://github.com/w3c/web-platform-tests/pull/1173">https://github.com/w3c/web-platform-tests/pull/1173</a></li>
 <li id="l-157"><a href="#l-157">#</a> [14:08] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-158"><a href="#l-158">#</a> [14:09] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-159"><a href="#l-159">#</a> [14:09] -gitbot:#testing- [web-platform-tests] dontcallmedom closed pull request #160: Test navigator.getUserMedia success callback is triggered when user allo... (master...submissions/kylebarrow) <a href="https://github.com/w3c/web-platform-tests/pull/160">https://github.com/w3c/web-platform-tests/pull/160</a></li>
 <li id="l-160"><a href="#l-160">#</a> [14:09] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-161" _a="Ms2ger"><a href="#l-161">#</a> [14:09] &lt;Ms2ger&gt; dom++</li>
 <li id="l-162" _a="Ms2ger"><a href="#l-162">#</a> [14:09] &lt;Ms2ger&gt; Oh look, I missed that we passed 1000 closed PRs</li>
 <li id="l-163" _a="plh" _r="so"><a href="#l-163">#</a> [14:10] &lt;plh&gt; so, for 1077, I'm thinking of two things. either I retire the PR and submit one to just do the example.org comment change, and do a separate one to do fixes for dnd tests</li>
 <li id="l-164"><a href="#l-164">#</a> [14:10] * Quits: Lachy (~Lachy@public.cloak) (Client closed connection)</li>
 <li id="l-165"><a href="#l-165">#</a> [14:10] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-166" _a="plh"><a href="#l-166">#</a> [14:10] &lt;plh&gt; look into how to disable entirely the https tests to make them fail fast</li>
 <li id="l-167" _a="plh"><a href="#l-167">#</a> [14:10] &lt;plh&gt; s//or/</li>
 <li id="l-168"><a href="#l-168">#</a> [14:25] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-169"><a href="#l-169">#</a> [14:25] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #1175: Cleanup the selectors-api submissions. (master...selectors2-cleanup) <a href="https://github.com/w3c/web-platform-tests/pull/1175">https://github.com/w3c/web-platform-tests/pull/1175</a></li>
 <li id="l-170"><a href="#l-170">#</a> [14:25] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-171" _a="Ms2ger" _r="MikeSmith"><a href="#l-171">#</a> [14:26] &lt;Ms2ger&gt; MikeSmith, if you're still in the mood... :)</li>
 <li id="l-172" _a="MikeSmith" _r="Ms2ger"><a href="#l-172">#</a> [14:37] &lt;MikeSmith&gt; Ms2ger: scribing atm :(</li>
 <li id="l-173" _a="MikeSmith"><a href="#l-173">#</a> [14:37] &lt;MikeSmith&gt; can do in 30 minutes or so</li>
 <li id="l-174"><a href="#l-174">#</a> [14:42] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-175"><a href="#l-175">#</a> [14:49] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-176"><a href="#l-176">#</a> [14:51] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-177"><a href="#l-177">#</a> [14:51] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #1175: Cleanup the selectors-api submissions. (master...selectors2-cleanup) <a href="https://github.com/w3c/web-platform-tests/pull/1175">https://github.com/w3c/web-platform-tests/pull/1175</a></li>
 <li id="l-178"><a href="#l-178">#</a> [14:51] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-179"><a href="#l-179">#</a> [14:54] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-180" _a="MikeSmith"><a href="#l-180">#</a> [15:06] &lt;MikeSmith&gt; scribing completed</li>
 <li id="l-181" _a="MikeSmith"><a href="#l-181">#</a> [15:06] &lt;MikeSmith&gt; damn but I see jgraham beat me to it in the mean time</li>
 <li id="l-182" _a="MikeSmith"><a href="#l-182">#</a> [15:06] &lt;MikeSmith&gt; you guys are just doing too much leading by example</li>
 <li id="l-183" _a="MikeSmith"><a href="#l-183">#</a> [15:07] &lt;MikeSmith&gt; forget the army of the 300</li>
 <li id="l-184" _a="jgraham" _r="MikeSmith"><a href="#l-184">#</a> [15:08] &lt;jgraham&gt; MikeSmith: Sorry. Maybe we can organise a telecon so we can resolve to take an action item to eventually do some review?</li>
 <li id="l-185" _a="MikeSmith"><a href="#l-185">#</a> [15:08] &lt;MikeSmith&gt; wpt has the army of the 3 (or so)</li>
 <li id="l-186" _a="MikeSmith" _r="jgraham"><a href="#l-186">#</a> [15:08] &lt;MikeSmith&gt; jgraham: I've pencilled that in to my Blackberry</li>
 <li id="l-187" _a="plh"><a href="#l-187">#</a> [15:09] &lt;plh&gt; hold on, what about the 8 months advance notice before doing a telcon?</li>
 <li id="l-188" _a="MikeSmith" _r="plh"><a href="#l-188">#</a> [15:13] &lt;MikeSmith&gt; plh: I asked the AC about that and 19.4 people responded to say that I didn't need to, so I'm claiming &quot;overwhelming consensus&quot;</li>
 <li id="l-189" _a="MikeSmith"><a href="#l-189">#</a> [15:15] &lt;MikeSmith&gt; damn , down to only 116 open PRs. This is getting scary</li>
 <li id="l-190" _a="MikeSmith"><a href="#l-190">#</a> [15:15] &lt;MikeSmith&gt; oh and in the mean time, wpt crossed the 1000 closed PRs milestone</li>
 <li id="l-191" _a="ato"><a href="#l-191">#</a> [15:42] &lt;ato&gt; When developing testharness.js TCs, is the usual approach to keep a process of serve.py running in the background?</li>
 <li id="l-192" _a="Ms2ger"><a href="#l-192">#</a> [15:42] &lt;Ms2ger&gt; I often do</li>
 <li id="l-193" _a="ato"><a href="#l-193">#</a> [15:45] &lt;ato&gt; Thanks</li>
 <li id="l-194" _a="Ms2ger" _r="zcorpan"><a href="#l-194">#</a> [15:53] &lt;Ms2ger&gt; zcorpan, might <a href="https://critic.hoppipolla.co.uk/r/2322">https://critic.hoppipolla.co.uk/r/2322</a> be of interest to you?</li>
 <li id="l-195"><a href="#l-195">#</a> [15:54] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-196" _a="zcorpan"><a href="#l-196">#</a> [15:57] &lt;zcorpan&gt; yep</li>
 <li id="l-197" _a="zcorpan"><a href="#l-197">#</a> [15:58] &lt;zcorpan&gt; what was the thing with characterSet and casing?</li>
 <li id="l-198" _a="heycam"><a href="#l-198">#</a> [15:58] * heycam is now known as heycam|away</li>
 <li id="l-199"><a href="#l-199">#</a> [15:59] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-200"><a href="#l-200">#</a> [15:59] -gitbot:#testing- [web-platform-tests] jgraham created jgraham/html5lib_timeout (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/c1c8fe2fa517">https://github.com/w3c/web-platform-tests/commit/c1c8fe2fa517</a></li>
 <li id="l-201"><a href="#l-201">#</a> [15:59] -gitbot:#testing- web-platform-tests/jgraham/html5lib_timeout c1c8fe2 James Graham: Give html5lib tests a long timeout as they can be slow in debug builds.</li>
 <li id="l-202"><a href="#l-202">#</a> [15:59] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-203"><a href="#l-203">#</a> [15:59] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-204"><a href="#l-204">#</a> [15:59] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #1176: Give html5lib tests a long timeout as they can be slow in debug builds. (master...jgraham/html5lib_timeout) <a href="https://github.com/w3c/web-platform-tests/pull/1176">https://github.com/w3c/web-platform-tests/pull/1176</a></li>
 <li id="l-205"><a href="#l-205">#</a> [15:59] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-206"><a href="#l-206">#</a> [15:59] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-207"><a href="#l-207">#</a> [15:59] -gitbot:#testing- [web-platform-tests] zcorpan closed pull request #1166: Test the metadata used for createHTMLDocument(). (master...createHTMLDocument) <a href="https://github.com/w3c/web-platform-tests/pull/1166">https://github.com/w3c/web-platform-tests/pull/1166</a></li>
 <li id="l-208"><a href="#l-208">#</a> [15:59] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-209" _a="jgraham" _r="Ms2ger"><a href="#l-209">#</a> [16:00] &lt;jgraham&gt; Ms2ger: <a href="https://critic.hoppipolla.co.uk/r/2353">https://critic.hoppipolla.co.uk/r/2353</a> r?</li>
 <li id="l-210" _a="Ms2ger"><a href="#l-210">#</a> [16:01] &lt;Ms2ger&gt; r+</li>
 <li id="l-211" _a="jgraham"><a href="#l-211">#</a> [16:02] &lt;jgraham&gt; Thanks</li>
 <li id="l-212"><a href="#l-212">#</a> [16:02] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-213"><a href="#l-213">#</a> [16:02] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #1176: Give html5lib tests a long timeout as they can be slow in debug builds. (master...jgraham/html5lib_timeout) <a href="https://github.com/w3c/web-platform-tests/pull/1176">https://github.com/w3c/web-platform-tests/pull/1176</a></li>
 <li id="l-214"><a href="#l-214">#</a> [16:02] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-215"><a href="#l-215">#</a> [16:02] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-216"><a href="#l-216">#</a> [16:02] -gitbot:#testing- [web-platform-tests] jgraham pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/cf45429d2ae26bd137cdfd400202fbaa4ba4cd9d">https://github.com/w3c/web-platform-tests/commit/cf45429d2ae26bd137cdfd400202fbaa4ba4cd9d</a></li>
 <li id="l-217"><a href="#l-217">#</a> [16:02] -gitbot:#testing- web-platform-tests/master cf45429 jgraham: Merge pull request #1176 from w3c/jgraham/html5lib_timeout...</li>
 <li id="l-218"><a href="#l-218">#</a> [16:02] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-219" _a="ato" _r="jgraham"><a href="#l-219">#</a> [16:04] &lt;ato&gt; jgraham: I see testharness.js uses a function called expose which moves stuff onto the window object.  Is this the ideal way to make functions from a closure public?</li>
 <li id="l-220" _a="jgraham" _r="ato"><a href="#l-220">#</a> [16:04] &lt;jgraham&gt; ato: Probably not</li>
 <li id="l-221" _a="ato"><a href="#l-221">#</a> [16:04] &lt;ato&gt; I mean the alternative is that I return a dict from the closure</li>
 <li id="l-222"><a href="#l-222">#</a> [16:05] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-223"><a href="#l-223">#</a> [16:05] -gitbot:#testing- [web-platform-tests] Ms2ger deleted jgraham/html5lib_timeout at c1c8fe2: <a href="https://github.com/w3c/web-platform-tests/commit/c1c8fe2">https://github.com/w3c/web-platform-tests/commit/c1c8fe2</a></li>
 <li id="l-224"><a href="#l-224">#</a> [16:05] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-225" _a="Ms2ger" _r="ato"><a href="#l-225">#</a> [16:05] &lt;Ms2ger&gt; ato, no, the alternative is window.foo = ...</li>
 <li id="l-226" _a="ato"><a href="#l-226">#</a> [16:05] &lt;ato&gt; I'm not asking to interrogate, but because I have no idea what I'm doing.</li>
 <li id="l-227" _a="jgraham" _r="Yeah"><a href="#l-227">#</a> [16:05] &lt;jgraham&gt; Yeah, but then the caller has to know how to handle that dict, which is generally OK, but not a great idea for this case</li>
 <li id="l-228" _a="ato" _r="Right"><a href="#l-228">#</a> [16:05] &lt;ato&gt; Right, so my use case is that I need to define a bunch of helper functions which I /don't/ want to expose, and the public functions needs to have access to them.</li>
 <li id="l-229" _a="ato"><a href="#l-229">#</a> [16:06] &lt;ato&gt; Oh right, because you're exposing individual functions and not methods on an object.</li>
 <li id="l-230" _a="jgraham"><a href="#l-230">#</a> [16:06] &lt;jgraham&gt; In general not putting stuff on the global at all is better; just return an object with the public functions</li>
 <li id="l-231" _a="jgraham"><a href="#l-231">#</a> [16:07] &lt;jgraham&gt; For th.js making people do testharness.assert_equals (or even t.assert_equals) seemed unreasonable</li>
 <li id="l-232" _a="ato"><a href="#l-232">#</a> [16:07] &lt;ato&gt; Absolutely.</li>
 <li id="l-233" _a="Ms2ger"><a href="#l-233">#</a> [16:08] &lt;Ms2ger&gt; Could've made it t.equals</li>
 <li id="l-234" _a="ato"><a href="#l-234">#</a> [16:08] &lt;ato&gt; I'm just trying to suck up some JS patterns.</li>
 <li id="l-235" _a="Ms2ger" _r="ato"><a href="#l-235">#</a> [16:08] &lt;Ms2ger&gt; ato, my poor soul :)</li>
 <li id="l-236" _a="ato"><a href="#l-236">#</a> [16:08] &lt;ato&gt; You're not helping!</li>
 <li id="l-237" _a="Ms2ger"><a href="#l-237">#</a> [16:08] &lt;Ms2ger&gt; When am I?</li>
 <li id="l-238" _a="jgraham" _r="Ms2ger"><a href="#l-238">#</a> [16:09] &lt;jgraham&gt; Ms2ger: Yeah or assert.equals or ass.equals…</li>
 <li id="l-239" _a="Ms2ger"><a href="#l-239">#</a> [16:10] &lt;Ms2ger&gt; Many options</li>
 <li id="l-240" _a="ato"><a href="#l-240">#</a> [16:10] &lt;ato&gt; ass.own_property(…) would've been hillarious.</li>
 <li id="l-241"><a href="#l-241">#</a> [16:38] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-242"><a href="#l-242">#</a> [16:59] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-243"><a href="#l-243">#</a> [16:59] -gitbot:#testing- [web-platform-tests] jgraham created jgraham/whitespace/trailing/2dcontext (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/fb81d51eb858">https://github.com/w3c/web-platform-tests/commit/fb81d51eb858</a></li>
 <li id="l-244"><a href="#l-244">#</a> [16:59] -gitbot:#testing- web-platform-tests/jgraham/whitespace/trailing/2dcontext fb81d51 James Graham: 2dcontext cleanup trailing whitespace</li>
 <li id="l-245"><a href="#l-245">#</a> [16:59] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-246"><a href="#l-246">#</a> [16:59] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-247"><a href="#l-247">#</a> [16:59] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #1177: 2dcontext cleanup trailing whitespace (master...jgraham/whitespace/trailing/2dcontext) <a href="https://github.com/w3c/web-platform-tests/pull/1177">https://github.com/w3c/web-platform-tests/pull/1177</a></li>
 <li id="l-248"><a href="#l-248">#</a> [16:59] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-249" _a="jgraham"><a href="#l-249">#</a> [17:02] &lt;jgraham&gt; did travis always email in response to wpt PRs?</li>
 <li id="l-250" _a="MikeSmith" _r="jgraham"><a href="#l-250">#</a> [17:05] &lt;MikeSmith&gt; jgraham: I think it does each time your commit a change to the PR (when you create the PR and for each change after)</li>
 <li id="l-251"><a href="#l-251">#</a> [17:05] * Quits: MikeSmith (~MikeSmith@public.cloak) (&quot;leaving&quot;)</li>
 <li id="l-252"><a href="#l-252">#</a> [17:06] * Joins: MikeSmith (~MikeSmith@public.cloak)</li>
 <li id="l-253"><a href="#l-253">#</a> [17:26] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-254"><a href="#l-254">#</a> [17:30] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Client closed connection)</li>
 <li id="l-255"><a href="#l-255">#</a> [17:36] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-256"><a href="#l-256">#</a> [17:41] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-257"><a href="#l-257">#</a> [17:48] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-258"><a href="#l-258">#</a> [17:48] -gitbot:#testing- [web-platform-tests] jgraham created jgraham/whitespace/trailing/html (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/99d5ba70d2f4">https://github.com/w3c/web-platform-tests/commit/99d5ba70d2f4</a></li>
 <li id="l-259"><a href="#l-259">#</a> [17:48] -gitbot:#testing- web-platform-tests/jgraham/whitespace/trailing/html 99d5ba7 James Graham: Clean up trailing whitespace under html/</li>
 <li id="l-260"><a href="#l-260">#</a> [17:48] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-261"><a href="#l-261">#</a> [17:48] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-262"><a href="#l-262">#</a> [17:48] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #1178: Clean up trailing whitespace under html/ (master...jgraham/whitespace/trailing/html) <a href="https://github.com/w3c/web-platform-tests/pull/1178">https://github.com/w3c/web-platform-tests/pull/1178</a></li>
 <li id="l-263"><a href="#l-263">#</a> [17:48] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-264"><a href="#l-264">#</a> [17:56] * Quits: lmclister (~lmclister@public.cloak) (Client closed connection)</li>
 <li id="l-265"><a href="#l-265">#</a> [17:57] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-266"><a href="#l-266">#</a> [17:58] * Joins: lmcliste_ (~lmclister@public.cloak)</li>
 <li id="l-267"><a href="#l-267">#</a> [17:58] * Quits: lmclister (~lmclister@public.cloak) (Client closed connection)</li>
 <li id="l-268"><a href="#l-268">#</a> [17:59] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-269"><a href="#l-269">#</a> [18:15] * Quits: dom (dom@public.cloak) (&quot;&quot;)</li>
 <li id="l-270"><a href="#l-270">#</a> [18:39] * Quits: bret (~bret@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-271"><a href="#l-271">#</a> [18:46] * Joins: bret (~bret@public.cloak)</li>
 <li id="l-272"><a href="#l-272">#</a> [19:05] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-273"><a href="#l-273">#</a> [19:05] * Quits: lmcliste_ (~lmclister@public.cloak) (Client closed connection)</li>
 <li id="l-274"><a href="#l-274">#</a> [19:19] * Joins: lmcliste_ (~lmclister@public.cloak)</li>
 <li id="l-275"><a href="#l-275">#</a> [19:19] * Quits: lmclister (~lmclister@public.cloak) (Client closed connection)</li>
 <li id="l-276"><a href="#l-276">#</a> [19:34] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-277"><a href="#l-277">#</a> [19:34] * Quits: lmcliste_ (~lmclister@public.cloak) (Client closed connection)</li>
 <li id="l-278"><a href="#l-278">#</a> [20:11] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-279"><a href="#l-279">#</a> [20:12] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-280"><a href="#l-280">#</a> [20:16] * Joins: lmcliste_ (~lmclister@public.cloak)</li>
 <li id="l-281"><a href="#l-281">#</a> [20:16] * Quits: lmclister (~lmclister@public.cloak) (Client closed connection)</li>
 <li id="l-282"><a href="#l-282">#</a> [20:35] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-283"><a href="#l-283">#</a> [20:35] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #1179: Update and improve the test for WorkerNavigator properties. (master...navigator) <a href="https://github.com/w3c/web-platform-tests/pull/1179">https://github.com/w3c/web-platform-tests/pull/1179</a></li>
 <li id="l-284"><a href="#l-284">#</a> [20:35] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-285"><a href="#l-285">#</a> [20:36] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-286"><a href="#l-286">#</a> [20:36] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #1177: 2dcontext cleanup trailing whitespace (master...jgraham/whitespace/trailing/2dcontext) <a href="https://github.com/w3c/web-platform-tests/pull/1177">https://github.com/w3c/web-platform-tests/pull/1177</a></li>
 <li id="l-287"><a href="#l-287">#</a> [20:36] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-288"><a href="#l-288">#</a> [20:36] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-289"><a href="#l-289">#</a> [20:36] -gitbot:#testing- [web-platform-tests] Ms2ger pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/88d37e085cab83fffa63b734379287f5219a0643">https://github.com/w3c/web-platform-tests/commit/88d37e085cab83fffa63b734379287f5219a0643</a></li>
 <li id="l-290"><a href="#l-290">#</a> [20:36] -gitbot:#testing- web-platform-tests/master 88d37e0 Ms2ger: Merge pull request #1177 from w3c/jgraham/whitespace/trailing/2dcontext...</li>
 <li id="l-291"><a href="#l-291">#</a> [20:36] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-292"><a href="#l-292">#</a> [20:36] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-293"><a href="#l-293">#</a> [20:36] -gitbot:#testing- [web-platform-tests] Ms2ger deleted jgraham/whitespace/trailing/2dcontext at fb81d51: <a href="https://github.com/w3c/web-platform-tests/commit/fb81d51">https://github.com/w3c/web-platform-tests/commit/fb81d51</a></li>
 <li id="l-294"><a href="#l-294">#</a> [20:36] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-295"><a href="#l-295">#</a> [20:38] * Quits: lmcliste_ (~lmclister@public.cloak) (&quot;&quot;)</li>
 <li id="l-296"><a href="#l-296">#</a> [20:38] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-297" _a="Ms2ger" _r="jgraham"><a href="#l-297">#</a> [20:38] &lt;Ms2ger&gt; jgraham, r? <a href="https://critic.hoppipolla.co.uk/r/2359">https://critic.hoppipolla.co.uk/r/2359</a></li>
 <li id="l-298"><a href="#l-298">#</a> [20:39] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-299" _a="Ms2ger"><a href="#l-299">#</a> [20:42] &lt;Ms2ger&gt; Or zcorpan, r? <a href="https://critic.hoppipolla.co.uk/r/2359">https://critic.hoppipolla.co.uk/r/2359</a></li>
 <li id="l-300"><a href="#l-300">#</a> [20:46] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-301" _a="Ms2ger"><a href="#l-301">#</a> [20:48] &lt;Ms2ger&gt; Ha</li>
 <li id="l-302"><a href="#l-302">#</a> [20:50] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-303"><a href="#l-303">#</a> [20:50] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #1180: Cleanup navigator_online_event-manual.html. (master...online-event) <a href="https://github.com/w3c/web-platform-tests/pull/1180">https://github.com/w3c/web-platform-tests/pull/1180</a></li>
 <li id="l-304"><a href="#l-304">#</a> [20:50] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-305"><a href="#l-305">#</a> [20:55] * Joins: glenn_ (~gadams@public.cloak)</li>
 <li id="l-306"><a href="#l-306">#</a> [21:00] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-307"><a href="#l-307">#</a> [21:19] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-308"><a href="#l-308">#</a> [21:19] -gitbot:#testing- [web-platform-tests] chaals opened pull request #1181: Create implementation report for CR (master...master) <a href="https://github.com/w3c/web-platform-tests/pull/1181">https://github.com/w3c/web-platform-tests/pull/1181</a></li>
 <li id="l-309"><a href="#l-309">#</a> [21:19] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-310"><a href="#l-310">#</a> [21:21] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-311"><a href="#l-311">#</a> [21:21] -gitbot:#testing- [web-platform-tests] chaals closed pull request #1181: Create implementation report for CR (master...master) <a href="https://github.com/w3c/web-platform-tests/pull/1181">https://github.com/w3c/web-platform-tests/pull/1181</a></li>
 <li id="l-312"><a href="#l-312">#</a> [21:21] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-313"><a href="#l-313">#</a> [21:22] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-314"><a href="#l-314">#</a> [21:22] -gitbot:#testing- [web-platform-tests] chaals reopened pull request #1181: Create implementation report for CR (master...master) <a href="https://github.com/w3c/web-platform-tests/pull/1181">https://github.com/w3c/web-platform-tests/pull/1181</a></li>
 <li id="l-315"><a href="#l-315">#</a> [21:22] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-316"><a href="#l-316">#</a> [21:23] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-317"><a href="#l-317">#</a> [21:23] -gitbot:#testing- [web-platform-tests] liamquin pushed 2 new commits to master: <a href="https://github.com/w3c/web-platform-tests/compare/88d37e085cab...c7d00231cf5b">https://github.com/w3c/web-platform-tests/compare/88d37e085cab...c7d00231cf5b</a></li>
 <li id="l-318"><a href="#l-318">#</a> [21:23] -gitbot:#testing- web-platform-tests/master 56b0dec chaals: Create implementation report for CR...</li>
 <li id="l-319"><a href="#l-319">#</a> [21:23] -gitbot:#testing- web-platform-tests/master c7d0023 Liam Quin: Merge pull request #1181 from chaals/master...</li>
 <li id="l-320"><a href="#l-320">#</a> [21:23] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-321"><a href="#l-321">#</a> [21:23] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-322"><a href="#l-322">#</a> [21:23] -gitbot:#testing- [web-platform-tests] liamquin closed pull request #1181: Create implementation report for CR (master...master) <a href="https://github.com/w3c/web-platform-tests/pull/1181">https://github.com/w3c/web-platform-tests/pull/1181</a></li>
 <li id="l-323"><a href="#l-323">#</a> [21:23] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-324"><a href="#l-324">#</a> [21:32] * Joins: zcorpan_ (~zcorpan@public.cloak)</li>
 <li id="l-325"><a href="#l-325">#</a> [21:36] * Quits: zcorpan_ (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-326"><a href="#l-326">#</a> [21:44] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-327"><a href="#l-327">#</a> [21:44] -gitbot:#testing- [web-platform-tests] chaals opened pull request #1182: in progress... (master...master) <a href="https://github.com/w3c/web-platform-tests/pull/1182">https://github.com/w3c/web-platform-tests/pull/1182</a></li>
 <li id="l-328"><a href="#l-328">#</a> [21:44] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-329"><a href="#l-329">#</a> [21:46] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-330"><a href="#l-330">#</a> [21:46] -gitbot:#testing- [web-platform-tests] liamquin pushed 3 new commits to master: <a href="https://github.com/w3c/web-platform-tests/compare/c7d00231cf5b...b9e72129df5a">https://github.com/w3c/web-platform-tests/compare/c7d00231cf5b...b9e72129df5a</a></li>
 <li id="l-331"><a href="#l-331">#</a> [21:46] -gitbot:#testing- web-platform-tests/master ebb56ec chaals: start editing...</li>
 <li id="l-332"><a href="#l-332">#</a> [21:46] -gitbot:#testing- web-platform-tests/master 87fa03f chaals: cleaning and explaining...</li>
 <li id="l-333"><a href="#l-333">#</a> [21:46] -gitbot:#testing- web-platform-tests/master b9e7212 Liam Quin: Merge pull request #1182 from chaals/master...</li>
 <li id="l-334"><a href="#l-334">#</a> [21:46] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-335" _a="plh"><a href="#l-335">#</a> [21:46] &lt;plh&gt; hu.... implementation report in WPT?!?!</li>
 <li id="l-336" _a="plh"><a href="#l-336">#</a> [21:46] * plh goes hunt Liam and chaals</li>
 <li id="l-337" _a="Ms2ger" _r="Uh"><a href="#l-337">#</a> [21:48] &lt;Ms2ger&gt; Uh, out with that</li>
 <li id="l-338" _a="plh"><a href="#l-338">#</a> [21:49] &lt;plh&gt; I caught them and told them to remove it</li>
 <li id="l-339" _a="plh"><a href="#l-339">#</a> [21:49] &lt;plh&gt; what's the best to undo here?</li>
 <li id="l-340" _a="plh"><a href="#l-340">#</a> [21:49] &lt;plh&gt; s//way/</li>
 <li id="l-341" _a="Ms2ger"><a href="#l-341">#</a> [21:49] &lt;Ms2ger&gt; I think there's a revert button on the PR</li>
 <li id="l-342" _a="plh"><a href="#l-342">#</a> [21:54] &lt;plh&gt; used it but no luck, looks like a new PR is needed to undo</li>
 <li id="l-343" _a="plh"><a href="#l-343">#</a> [21:55] &lt;plh&gt; I'm giving them a chance to undo this themselves</li>
 <li id="l-344" _a="plh"><a href="#l-344">#</a> [21:56] &lt;plh&gt; in the meantime, I made some progress on <a href="https://github.com/w3c/web-platform-tests/pull/1077">https://github.com/w3c/web-platform-tests/pull/1077</a></li>
 <li id="l-345"><a href="#l-345">#</a> [22:02] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-346"><a href="#l-346">#</a> [22:13] * Joins: ptressel (~chatzilla@public.cloak)</li>
 <li id="l-347" _a="jgraham"><a href="#l-347">#</a> [22:16] &lt;jgraham&gt; Yeah we just need a new PR that rms the added file</li>
 <li id="l-348"><a href="#l-348">#</a> [22:23] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-349"><a href="#l-349">#</a> [22:27] * Quits: glenn_ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-350"><a href="#l-350">#</a> [22:27] * Joins: glenn__ (~gadams@public.cloak)</li>
 <li id="l-351"><a href="#l-351">#</a> [22:32] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-352"><a href="#l-352">#</a> [22:33] * Joins: lmcliste_ (~lmclister@public.cloak)</li>
 <li id="l-353"><a href="#l-353">#</a> [22:33] * Quits: lmclister (~lmclister@public.cloak) (Client closed connection)</li>
 <li id="l-354"><a href="#l-354">#</a> [22:34] * Quits: lmcliste_ (~lmclister@public.cloak) (Client closed connection)</li>
 <li id="l-355"><a href="#l-355">#</a> [22:34] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-356"><a href="#l-356">#</a> [22:37] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-357"><a href="#l-357">#</a> [22:37] -gitbot:#testing- [web-platform-tests] Ms2ger pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/09584d7894ab973097359dcda93ac9dfb1eb1f11">https://github.com/w3c/web-platform-tests/commit/09584d7894ab973097359dcda93ac9dfb1eb1f11</a></li>
 <li id="l-358"><a href="#l-358">#</a> [22:37] -gitbot:#testing- web-platform-tests/master 09584d7 Ms2ger: Remove implementation reports....</li>
 <li id="l-359"><a href="#l-359">#</a> [22:37] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-360"><a href="#l-360">#</a> [22:50] * Joins: mdas_ (~mdas@public.cloak)</li>
 <li id="l-361"><a href="#l-361">#</a> [22:50] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-362"><a href="#l-362">#</a> [23:10] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-363"><a href="#l-363">#</a> [23:10] * Quits: lmclister (~lmclister@public.cloak) (Client closed connection)</li>
 <li id="l-364"><a href="#l-364">#</a> [23:11] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-365"><a href="#l-365">#</a> [23:15] * Quits: glenn__ (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-366"><a href="#l-366">#</a> [23:23] * Quits: mdas_ (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-367"><a href="#l-367">#</a> [23:24] * Joins: lmcliste_ (~lmclister@public.cloak)</li>
 <li id="l-368"><a href="#l-368">#</a> [23:24] * Quits: lmclister (~lmclister@public.cloak) (Client closed connection)</li>
 <li id="l-369"><a href="#l-369">#</a> [23:46] * Quits: Ms2ger (~Ms2ger@public.cloak) (&quot;nn&quot;)</li>
 <li id="l-370"><a href="#l-370">#</a> Session Close: Fri Aug 15 00:00:00 2014</li>
</ol>
<p id="end">The end :)
